Fixed a potential security break
When rounding position to 64 as the counter to XORKeyStream and the
writing blocks smaller than 64, we have a nonce reuse.

When calling Write(), I was passing cw.position / 64 as the counter to XORKeyStream. This was because internally Salsa20 operates on 64 byte blocks and using an internal counter to track the number of blocks encrypted. I wanted to keep the two counters consistent, and maximize the size of file that could be encrypted, so I figured I should pass a counter that also tracks the number of 64 byte blocks encrypted. This turned out to be a mistake:

Starting with counter = 0
cw.Write([64]byte) (64/64 = 1, 0+1 = 1,so the next call will have an internal
counter of 1)
cw.Write([64]byte) (64/64 = 1, 1+1 = 2, so the next call will have an
internal counter of 2)

This will be secure if every write we have is cleanly divisible to 64
byte chunks, as the counter will increment properly every time and will
correctly count the number of 64 byte chunks that have been encrypted.

However, if we start to write slices that aren't cleanly divisible to 64
byte chunks, we have an issue:

Starting with counter = 0
cw.Write([8]byte) (8/64 = 0, 0+0 = 0, next call will have an internal
counter of 0)
cw.Write([8]byte) (8/64 = 0, 0+0 = 0, next call will have an internal
counter of 0)

We've reused the same counter to encrypt two different pieces of data.
If an attacker knows the plaintext of either of those 8 byte slice,
they can XOR the plaintext with the 8 byte ciphertext to get the
keystream for that chunk. They can then XOR that keystream with the next
8 byte ciphertext to get the plaintext for that slice. We've lost
confidentiality for that chunk.

There's a couple of different solutions here.

1. If we have an incomplete block, we can save it somewhere and wait for
the next call to Write() to fill up the incomplete block, and then write
that to the file.

cw.Write([8]byte) (8 < 64, so we save it in a temporary buffer)
cw.Write([64]byte) (8+64 > 64, so we add the first 56 bytes to the
temporary buffer and then encrypt & write that out, then save the
remaining 8 bytes into the temporary buffer). The downside to this
method is a potential performance hit from having to check the temporary
buffer every write.

2. We can directly pass cw.position to XORKeyStream as the counter. This
is what I'm currently doing but I'll likely switch to option 1. The
downside to this is it limits the maximum filesize we could encrypt to (2^64 - 1)/64 =
256PiB which I believe is more than enough.

I chose the second method (for now) because I'm not yet sure of the
performance hit 1. might have, considering every write we'd have to
check the size of the temporary buffer and add values if we're
misaligned. I'll be doing some tests soon to evaluate the potential
performance impact and may change my solution based on that.